Work around a crash on re-attaching interfaces.
The test suite more or less reliably provokes a kernel crash when run on
xnu-2422.115.4. The reason is missing synchronization on terminating the
input thread for a network interface. If this doesn't complete before
the interface gets re-attached, ifnet_attach will run into failing
asserts. This code change adds extra synchronization in the kext to wait
for the input thread to terminate before returning from close() on the
character device.